Renaming files to fit project name s2s
