+++ test/hello_s2s.cpp
@@ -1,4 +1,4 @@
-#include "../include/struct_cast.hpp"
+#include "../include/s2s.hpp"
 #include <print>
 #include <expected>
 #include <fstream>
